🔧feat: 로그아웃시 필요한 토큰 검증 securityutils 에서 에러 검증 로직 추가

+++ backend/farmbti/src/main/java/com/backend/farmbti/auth/dto/SignUpRequest.java
@@ -17,7 +17,7 @@ public class SignUpRequest {
     private String password;
     private String name;
     private String address;
-    private Long gender;
+    private Byte gender;
     private LocalDate birth;
     private String profileImage;
